+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Pull Request: https://github.com/suzukiplan/vgssdk-pico/pull/6
 
 - `VGS::VDP::resize` メソッドを追加
+- vgsftv と vgslz4 コマンドの機能を vgsmml に統合
 
 ## Version 1.3.0